+++ types/winston-rollbar.d.ts
@@ -1,0 +1,1 @@
+declare module 'winston-transport-rollbar-3'